+++ crates/warpgate/src/plugin.rs
@@ -7,7 +7,6 @@ use once_map::OnceMap;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use starbase_styles::color;
-use std::collections::BTreeMap;
 use std::fmt::Debug;
 use std::path::{Path, PathBuf};
 use std::sync::{Arc, RwLock};
@@ -49,25 +48,6 @@ impl PluginContainer {
     /// Create a new container with the provided manifest.
     pub fn new_without_functions(id: Id, manifest: Manifest) -> miette::Result<PluginContainer> {
         Self::new(id, manifest, [])
-    }
-
-    /// Reload the plugin's configuration from the manifest.
-    pub fn reload_config(&mut self) -> miette::Result<()> {
-        // let config = self
-        //     .manifest
-        //     .config
-        //     .iter()
-        //     .map(|(k, v)| (k.to_owned(), Some(v.to_owned())))
-        //     .collect::<BTreeMap<_, _>>();
-
-        // TODO
-        // self.plugin
-        //     .write()
-        //     .expect("Failed to acquire write access!")
-        //     .set_config(&config)
-        //     .unwrap();
-
-        Ok(())
     }
 
     /// Call a function on the plugin with no input and cache the output before returning it.